fix(mcp): split inventory-items, add missing tests, fix formatting
- Split inventory-items.ts (216→56 lines) into read and write modules to
  stay under the 200-line oxlint limit; write tools live in
  inventory-items-write.ts (151 lines), barrel-exported via itemTools
- Add optNum helper to utils.ts
- Add explicit write-tool name assertions to index.test.ts (not just count)
- Add NOT_FOUND propagation test for connections.disconnect
- Fix formatting on test-helpers.ts and new split files via oxfmt
- Tighten delete location comment to a single line
- Mark PRD-103/104/105 as Done in epic status table
